fix: filters column with fixed size + delete some margin + mobile version
